third_party/blink/web_tests/wpt_internal/route/parsing/active-navigation-invalid.html

<!DOCTYPE html>
<title>Invalid :active-navigation() rules</title>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-navigation-1/#active-navigation-pseudo-class">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  test(()=> {
    test_invalid_rule(`:active-navigation { }`);
  }, "missing parentheses");

  test(()=> {
    test_invalid_rule(`:active-navigation(--rte garbage) { }`);
  }, "trailing garbage 1");

  test(()=> {
    test_invalid_rule(`:active-navigation(--rte --garbage) { }`);
  }, "trailing garbage 2");

  test(()=> {
    test_invalid_rule(`:active-navigation(url(/) with) { }`);
  }, "trailing garbage 3");
</script>
